pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# See https://setuptools.pypa.io/en/latest/userguide/quickstart.html for more project configuration options.
name = "flames"
dynamic = ["version"]
readme = "README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
authors = [
    {name = "Felipe Lopes de Oliveira", email = "felipe.lopes@nano.ufrj.br"}
]
requires-python = ">=3.10"
dependencies = [
  "pip",
  "jupyter",
  "jupyterlab",
  "numpy",
  "scipy",
  "ase",
  "pymatgen",
  "tqdm",
  "gemmi",
  "simplejson",
  "torch-dftd",
  "pymser"
]
license = {file = "LICENSE"}

[project.urls]
Homepage = "https://github.com/lipelopesoliveira/flames"
Repository = "https://github.com/lipelopesoliveira/flames"
Changelog = "https://github.com/lipelopesoliveira/flames/blob/main/CHANGELOG.md"
# Documentation = "https://flames.readthedocs.io/"

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy>=1.0,<2.0",
    "black>=23.0,<27.0",
    "isort>=5.12,<7.1",
    "pytest",
    "pytest-sphinx",
    "pytest-cov",
    "mace-torch==0.3.12",
    "twine>=1.11.0",
    "build",
    "setuptools",
    "wheel",
    "Sphinx>=6.0,<9.0",
    "furo==2025.12.19",
    "myst-parser>=1.0",
    "sphinx-copybutton",
    "sphinx-autobuild",
    "sphinx-autodoc-typehints==3.0.1",
    "packaging"
]

[tool.setuptools.packages.find]
exclude = [
    "*.tests",
    "*.tests.*",
    "tests.*",
    "tests",
    "docs*",
    "scripts*"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
flames = ["py.typed"]

[tool.setuptools.dynamic]
version = {attr = "flames.version.VERSION"}

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

# You can override these pyright settings by adding a personal pyrightconfig.json file.
[tool.pyright]
reportPrivateImportUsage = false

[tool.ruff]
line-length = 115
target-version = "py39"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
ignore_missing_imports = true
no_site_packages = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
strict_optional = false

[tool.pytest.ini_options]
testpaths = "tests/"
python_classes = [
  "Test*",
  "*Test"
]
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_level = "DEBUG"
